pyproject.toml
[tool.poetry]
name = "payments-robot"
version = "0.1.0"
description = ""
authors = ["Pablo Ganga <pabloganga@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<=3.13"
fastapi = "^0.111.1"
uvicorn = {extras = ["standard"], version = "^0.30.6"}
crewai = "^0.134.0"
python-dotenv = "^1.1.1"
aisuite = ">=0.1.10"
stripe = "^10.2.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
addopts = "-s"
